issue 26, renamed existing card components to reflect their purpose, created new base card component, changes to index and train stats to allow these changes to work properly
